Well, it compiles!  Let's see if it works...

+++ src/libes/ES_Synow_MOATSource.cc
@@ -56,15 +56,15 @@ ES::Synow::MOATSource::MOATSource( ES::Synow::Grid& grid, size_t const mu_size )
 
     // Allocate buffers needed for source metadata initialization.
 
-    _dev_v     = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( float  ) *           v_size               , NULL, NULL );
-    _dev_mu    = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( float  ) *           v_size * _mu_size * 2, NULL, NULL );
-    _dev_dmu   = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( float  ) *           v_size * _mu_size * 2, NULL, NULL );
-    _dev_shift = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( float  ) *           v_size * _mu_size * 2, NULL, NULL );
-    _dev_wl    = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( float  ) * wl_size                        , NULL, NULL );
-    _dev_tau   = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( float  ) * wl_size * v_size               , NULL, NULL );
+    _dev_v     = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( double ) *           v_size               , NULL, NULL );
+    _dev_mu    = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( double ) *           v_size * _mu_size * 2, NULL, NULL );
+    _dev_dmu   = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( double ) *           v_size * _mu_size * 2, NULL, NULL );
+    _dev_shift = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( double ) *           v_size * _mu_size * 2, NULL, NULL );
+    _dev_wl    = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( double ) * wl_size                        , NULL, NULL );
+    _dev_tau   = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( double ) * wl_size * v_size               , NULL, NULL );
     _dev_start = clCreateBuffer( _context, CL_MEM_READ_ONLY , sizeof( size_t ) * wl_size * v_size * _mu_size * 2, NULL, NULL );
-    _dev_in    = clCreateBuffer( _context, CL_MEM_READ_WRITE, sizeof( float  ) * wl_size * v_size * _mu_size * 2, NULL, NULL );
-    _dev_src   = clCreateBuffer( _context, CL_MEM_READ_WRITE, sizeof( float  ) * wl_size * v_size               , NULL, NULL );
+    _dev_in    = clCreateBuffer( _context, CL_MEM_READ_WRITE, sizeof( double ) * wl_size * v_size * _mu_size * 2, NULL, NULL );
+    _dev_src   = clCreateBuffer( _context, CL_MEM_READ_WRITE, sizeof( double ) * wl_size * v_size               , NULL, NULL );
 
     _global_work_size = v_size * _mu_size * 2;
     _local_work_size  = _mu_size * 2;
@@ -88,7 +88,7 @@ void ES::Synow::MOATSource::operator() ( const ES::Synow::Setup& setup )
 
     size_t i;
 
-    // Local caching.  TODO Cast?
+    // Local caching.
 
     double v_phot  = setup.v_phot;
     double v_outer = setup.v_outer;
@@ -102,15 +102,15 @@ void ES::Synow::MOATSource::operator() ( const ES::Synow::Setup& setup )
     for( size_t iv = 0; iv < v_size; ++ iv )
     {
 
-        float vv      = _grid->v[ iv ];
-        float mu_crit = iv > 0 ? sqrt( 1.0 - v_phot * v_phot / vv / vv ) : 0.0;
+        double vv      = _grid->v[ iv ];
+        double mu_crit = iv > 0 ? sqrt( 1.0 - v_phot * v_phot / vv / vv ) : 0.0;
 
-        float tmp_dmu;
-        float mu_init;
+        double tmp_dmu;
+        double mu_init;
 
         // Rays originating at the photosphere.
 
-        tmp_dmu = ( 1.0 - mu_crit ) / float( _mu_size );
+        tmp_dmu = ( 1.0 - mu_crit ) / double( _mu_size );
         mu_init = 1.0 - 0.5 * tmp_dmu;
 
         for( size_t im = 0; im < _mu_size; ++ im )
@@ -123,7 +123,7 @@ void ES::Synow::MOATSource::operator() ( const ES::Synow::Setup& setup )
 
         // Rays originating in the sky.
 
-        tmp_dmu = ( mu_crit + 1.0 ) / float( _mu_size );
+        tmp_dmu = ( mu_crit + 1.0 ) / double( _mu_size );
         mu_init = mu_crit - 0.5 * tmp_dmu;
 
         for( size_t im = 0; im < _mu_size; ++ im )
@@ -164,7 +164,7 @@ void ES::Synow::MOATSource::operator() ( const ES::Synow::Setup& setup )
             for( size_t im = 0; im < _mu_size; ++ im )
             {
                 _start[ i ] = 0;
-                float start_wl = _grid->wl[ iw ] * _shift[ j ];
+                double start_wl = _grid->wl[ iw ] * _shift[ j ];
                 while( _grid->wl[ _start[ i ] ] < start_wl ) ++ _start[ i ];
                 _in[ i ] = (*_grid->bb)( _grid->wl[ iw ] * _shift[ i ] ) * pow( _shift[ i ], 3 );
 
@@ -177,7 +177,7 @@ void ES::Synow::MOATSource::operator() ( const ES::Synow::Setup& setup )
             for( size_t im = 0; im < _mu_size; ++ im )
             {
                 _start[ i ] = 0;
-                float start_wl = _grid->wl[ iw ] * _shift[ j ];
+                double start_wl = _grid->wl[ iw ] * _shift[ j ];
                 while( _grid->wl[ _start[ i ] ] < start_wl ) ++ _start[ i ];
                 _in[ i ] = 0.0;
 
@@ -193,15 +193,15 @@ void ES::Synow::MOATSource::operator() ( const ES::Synow::Setup& setup )
 
     cl_event write_events[ 9 ];
 
-    clEnqueueWriteBuffer( _queue, _dev_v    , CL_TRUE, 0, sizeof( float  ) *           v_size               , (const void*)   _grid->v, 0, NULL, &(write_events[ 0 ]) );
-    clEnqueueWriteBuffer( _queue, _dev_mu   , CL_TRUE, 0, sizeof( float  ) *           v_size * _mu_size * 2, (const void*)        _mu, 0, NULL, &(write_events[ 1 ]) );
-    clEnqueueWriteBuffer( _queue, _dev_dmu  , CL_TRUE, 0, sizeof( float  ) *           v_size * _mu_size * 2, (const void*)       _dmu, 0, NULL, &(write_events[ 2 ]) );
-    clEnqueueWriteBuffer( _queue, _dev_shift, CL_TRUE, 0, sizeof( float  ) *           v_size * _mu_size * 2, (const void*)     _shift, 0, NULL, &(write_events[ 3 ]) );
-    clEnqueueWriteBuffer( _queue, _dev_wl   , CL_TRUE, 0, sizeof( float  ) * wl_used                        , (const void*)  _grid->wl, 0, NULL, &(write_events[ 4 ]) );
-    clEnqueueWriteBuffer( _queue, _dev_tau  , CL_TRUE, 0, sizeof( float  ) * wl_used * v_size               , (const void*) _grid->tau, 0, NULL, &(write_events[ 5 ]) );
+    clEnqueueWriteBuffer( _queue, _dev_v    , CL_TRUE, 0, sizeof( double ) *           v_size               , (const void*)   _grid->v, 0, NULL, &(write_events[ 0 ]) );
+    clEnqueueWriteBuffer( _queue, _dev_mu   , CL_TRUE, 0, sizeof( double ) *           v_size * _mu_size * 2, (const void*)        _mu, 0, NULL, &(write_events[ 1 ]) );
+    clEnqueueWriteBuffer( _queue, _dev_dmu  , CL_TRUE, 0, sizeof( double ) *           v_size * _mu_size * 2, (const void*)       _dmu, 0, NULL, &(write_events[ 2 ]) );
+    clEnqueueWriteBuffer( _queue, _dev_shift, CL_TRUE, 0, sizeof( double ) *           v_size * _mu_size * 2, (const void*)     _shift, 0, NULL, &(write_events[ 3 ]) );
+    clEnqueueWriteBuffer( _queue, _dev_wl   , CL_TRUE, 0, sizeof( double ) * wl_used                        , (const void*)  _grid->wl, 0, NULL, &(write_events[ 4 ]) );
+    clEnqueueWriteBuffer( _queue, _dev_tau  , CL_TRUE, 0, sizeof( double ) * wl_used * v_size               , (const void*) _grid->tau, 0, NULL, &(write_events[ 5 ]) );
     clEnqueueWriteBuffer( _queue, _dev_start, CL_TRUE, 0, sizeof( size_t ) * wl_used * v_size * _mu_size * 2, (const void*)     _start, 0, NULL, &(write_events[ 6 ]) );
-    clEnqueueWriteBuffer( _queue, _dev_in   , CL_TRUE, 0, sizeof( float  ) * wl_used * v_size * _mu_size * 2, (const void*)        _in, 0, NULL, &(write_events[ 7 ]) );
-    clEnqueueWriteBuffer( _queue, _dev_src  , CL_TRUE, 0, sizeof( float  ) * wl_used * v_size               , (const void*) _grid->src, 0, NULL, &(write_events[ 8 ]) );
+    clEnqueueWriteBuffer( _queue, _dev_in   , CL_TRUE, 0, sizeof( double ) * wl_used * v_size * _mu_size * 2, (const void*)        _in, 0, NULL, &(write_events[ 7 ]) );
+    clEnqueueWriteBuffer( _queue, _dev_src  , CL_TRUE, 0, sizeof( double ) * wl_used * v_size               , (const void*) _grid->src, 0, NULL, &(write_events[ 8 ]) );
 
     // Set kernel arguments.
 
@@ -212,16 +212,16 @@ void ES::Synow::MOATSource::operator() ( const ES::Synow::Setup& setup )
     clSetKernelArg( _kernel,  2, sizeof( size_t )                   , (void*) &_mu_size   );
     clSetKernelArg( _kernel,  3, sizeof( size_t )                   , (void*) &wl_used    );
     clSetKernelArg( _kernel,  4, sizeof( cl_mem )                   , (void*) &_dev_v     );
-    clSetKernelArg( _kernel,  5, sizeof( float  ) * v_size          ,         NULL        );
+    clSetKernelArg( _kernel,  5, sizeof( double ) * v_size          ,         NULL        );
     clSetKernelArg( _kernel,  6, sizeof( cl_mem )                   , (void*) &_dev_mu    );
     clSetKernelArg( _kernel,  7, sizeof( cl_mem )                   , (void*) &_dev_dmu   );
     clSetKernelArg( _kernel,  8, sizeof( cl_mem )                   , (void*) &_dev_shift );
     clSetKernelArg( _kernel,  9, sizeof( cl_mem )                   , (void*) &_dev_wl    );
-    clSetKernelArg( _kernel, 10, sizeof( float  ) * wl_used         ,         NULL        );
+    clSetKernelArg( _kernel, 10, sizeof( double ) * wl_used         ,         NULL        );
     clSetKernelArg( _kernel, 11, sizeof( cl_mem )                   , (void*) &_dev_tau   );
     clSetKernelArg( _kernel, 12, sizeof( cl_mem )                   , (void*) &_dev_start );
     clSetKernelArg( _kernel, 13, sizeof( cl_mem )                   , (void*) &_dev_in    );
-    clSetKernelArg( _kernel, 14, sizeof( float  ) * _local_work_size,         NULL        );
+    clSetKernelArg( _kernel, 14, sizeof( double ) * _local_work_size,         NULL        );
     clSetKernelArg( _kernel, 15, sizeof( cl_mem )                   , (void*) &_dev_src   );
 
     // Enqueue and execute kernel chain.
@@ -239,7 +239,7 @@ void ES::Synow::MOATSource::operator() ( const ES::Synow::Setup& setup )
     // Read back source function.
 
     cl_event read_event;
-    clEnqueueReadBuffer( _queue, _dev_src, CL_TRUE, 0, sizeof( float ) * wl_used * v_size, _grid->src, 1, &exec_event, &read_event ); // TODO to grid src
+    clEnqueueReadBuffer( _queue, _dev_src, CL_TRUE, 0, sizeof( double ) * wl_used * v_size, _grid->src, 1, &exec_event, &read_event );
     clWaitForEvents( 1, &read_event );
 
     // Cleanup.
